AuroraQ/requirements/base.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Async and Redis
asyncio
aioredis>=2.0.0
redis>=4.0.0

# Machine Learning
torch>=2.0.0
scikit-learn>=1.3.0

# Optimization
cvxpy>=1.3.0

# Data Collection
requests>=2.28.0
yfinance>=0.2.0

# Logging and Configuration
structlog>=22.0.0
pydantic>=2.0.0
PyYAML>=6.0

# Utilities
python-dotenv>=1.0.0
click>=8.0.0

# Optional: FinBERT sentiment analysis
transformers>=4.30.0
tokenizers>=0.13.0